Transition float to CGFloat, but also correct some int types to CGFloat where appropriate.
git-svn-id: svn://svn.savannah.nongnu.org/gap/trunk@3803 d25b4930-3292-e311-b074-0010c6de5da8

+++ system-apps/HelpViewer/TextFormatterXLP.h
@@ -71,7 +71,7 @@
 - (void) addImage: (NSImage*) img onString: (NSMutableAttributedString*) as;
 - (void) addLegendFig: (NSString*) imgpath withLegends: (NSArray*) plegends;
 - (void) addNote: (NSMutableAttributedString*) string withImage: (NSImage*) img withColor: (NSColor*) color;
-- (void) addRuleTo: (NSMutableAttributedString*) string withHeight: (int) height;
+- (void) addRuleTo: (NSMutableAttributedString*) string withHeight: (CGFloat) height;
 @end
 
 #endif
